# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Engineering Manual Knowledge Graph API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface UploadUrlRequest {
  name: string;
  size: number;
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
}

export type ManualStatus = typeof ManualStatus[keyof typeof ManualStatus];


export const ManualStatus = {
  pending: 'pending',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export interface Manual {
  id: number;
  name: string;
  filename: string;
  objectPath: string;
  status: ManualStatus;
  /** @nullable */
  processingPass?: number | null;
  /** @nullable */
  totalPages?: number | null;
  /** @nullable */
  documentType?: string | null;
  /** @nullable */
  errorMessage?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ManualInput {
  name: string;
  filename: string;
  objectPath: string;
}

export type GraphNodeType = typeof GraphNodeType[keyof typeof GraphNodeType];


export const GraphNodeType = {
  machine: 'machine',
  component: 'component',
  subsystem: 'subsystem',
  process: 'process',
  part: 'part',
  material: 'material',
  sensor: 'sensor',
  system: 'system',
  assembly: 'assembly',
  document_section: 'document_section',
} as const;

/**
 * @nullable
 */
export type GraphNodeProperties = { [key: string]: unknown } | null;

export interface GraphNode {
  id: number;
  manualId: number;
  name: string;
  type: GraphNodeType;
  description: string;
  /** @nullable */
  properties?: GraphNodeProperties;
  pageReferences?: number[];
  /** @nullable */
  orderIndex?: number | null;
  /** @nullable */
  manualName?: string | null;
}

export type GraphEdgeType = typeof GraphEdgeType[keyof typeof GraphEdgeType];


export const GraphEdgeType = {
  contains: 'contains',
  part_of: 'part_of',
  connects_to: 'connects_to',
  depends_on: 'depends_on',
  sequence: 'sequence',
  communicates_with: 'communicates_with',
  powers: 'powers',
  controls: 'controls',
  feeds_into: 'feeds_into',
  mounted_on: 'mounted_on',
} as const;

/**
 * @nullable
 */
export type GraphEdgeProperties = { [key: string]: unknown } | null;

export interface GraphEdge {
  id: number;
  manualId: number;
  sourceEntityId: number;
  targetEntityId: number;
  type: GraphEdgeType;
  label: string;
  /** @nullable */
  orderIndex?: number | null;
  /** @nullable */
  properties?: GraphEdgeProperties;
}

export interface GraphData {
  nodes: GraphNode[];
  edges: GraphEdge[];
}

export interface EntityDetail {
  entity: GraphNode;
  incomingEdges: GraphEdge[];
  outgoingEdges: GraphEdge[];
}

export type ManualStatsEntitiesByType = { [key: string]: unknown };

export type ManualStatsRelationshipsByType = { [key: string]: unknown };

export interface ManualStats {
  manualId: number;
  totalEntities: number;
  totalRelationships: number;
  entitiesByType: ManualStatsEntitiesByType;
  relationshipsByType: ManualStatsRelationshipsByType;
}

export type GlobalStatsEntitiesByType = { [key: string]: unknown };

export type GlobalStatsRelationshipsByType = { [key: string]: unknown };

export interface GlobalStats {
  totalManuals: number;
  totalEntities: number;
  totalRelationships: number;
  completedManuals: number;
  processingManuals: number;
  entitiesByType: GlobalStatsEntitiesByType;
  relationshipsByType: GlobalStatsRelationshipsByType;
}