# lib/api-zod/src/generated/types/graphData.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Engineering Manual Knowledge Graph API
 * OpenAPI spec version: 0.1.0
 */
import type { GraphEdge } from './graphEdge';
import type { GraphNode } from './graphNode';

export interface GraphData {
  nodes: GraphNode[];
  edges: GraphEdge[];
}