# lib/api-zod/src/generated/types/graphEdge.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Engineering Manual Knowledge Graph API
 * OpenAPI spec version: 0.1.0
 */
import type { GraphEdgeProperties } from './graphEdgeProperties';
import type { GraphEdgeType } from './graphEdgeType';

export interface GraphEdge {
  id: number;
  manualId: number;
  sourceEntityId: number;
  targetEntityId: number;
  type: GraphEdgeType;
  label: string;
  /** @nullable */
  orderIndex?: number | null;
  /** @nullable */
  properties?: GraphEdgeProperties;
}