# lib/api-zod/src/generated/types/graphNode.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Engineering Manual Knowledge Graph API
 * OpenAPI spec version: 0.1.0
 */
import type { GraphNodeProperties } from './graphNodeProperties';
import type { GraphNodeType } from './graphNodeType';

export interface GraphNode {
  id: number;
  manualId: number;
  name: string;
  type: GraphNodeType;
  description: string;
  /** @nullable */
  properties?: GraphNodeProperties;
  pageReferences?: number[];
  /** @nullable */
  orderIndex?: number | null;
  /** @nullable */
  manualName?: string | null;
}