# lib/api-zod/src/generated/types/manualInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Engineering Manual Knowledge Graph API
 * OpenAPI spec version: 0.1.0
 */

export interface ManualInput {
  name: string;
  filename: string;
  objectPath: string;
}