# lib/api-zod/src/generated/types/manualStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Engineering Manual Knowledge Graph API
 * OpenAPI spec version: 0.1.0
 */
import type { ManualStatsEntitiesByType } from './manualStatsEntitiesByType';
import type { ManualStatsRelationshipsByType } from './manualStatsRelationshipsByType';

export interface ManualStats {
  manualId: number;
  totalEntities: number;
  totalRelationships: number;
  entitiesByType: ManualStatsEntitiesByType;
  relationshipsByType: ManualStatsRelationshipsByType;
}